+++ src/index-home.js
@@ -1,18 +1,18 @@
 import './js/refs';
-import { uploadTrendMovies } from './js/createPageHome';
+import { uploadTrendMovies } from './js/create-page-home';
 import { topUp } from './js/top';
 // import './js/createPageHome';
 import './js/key-word-search';
-import './js/createSearchGallery';
+import './js/create-search-gallery';
 import './js/pagination';
 
-import './js/openMovieModal';
+// import './js/openMovieModal';
 
 // import './js/get-trend-movies';
 import './js/team-modal';
 import './js/dark-theme';
 // import './js/scroll';
-import topUp from './js/top';
+// import topUp from './js/top';
 
 uploadTrendMovies();
 topUp();
