+++ src/index.html
@@ -24,6 +24,6 @@
     
     <!-- <include src="./partials/pagination.html"></include> -->
     <include src="./partials/footer.html"></include>
-    <script type="module" src="./indexHome.js"></script>
+    <script type="module" src="./index-home.js"></script>
   </body>
 </html>
