+++ src/js/create-page-home.js
@@ -1,7 +1,7 @@
 import { refs } from './refs';
 import { pagination } from './pagination';
 import { createMovieCard } from './get-trend-movies';
-import { FilmsApiService } from './apiService';
+import { FilmsApiService } from './api-service';
 
 // import Pagination from 'tui-pagination';
