+++ src/js/create-search-gallery.js
