+++ src/js/scroll-blocker.js
