+++ src/js/team-modal.js
@@ -1,4 +1,4 @@
-import teamList from './teamList';
+import teamList from './team-list';
 const body = document.body;
 const list = document.querySelector('.team__list');
 const modal = document.querySelector('.goit-students');
