+++ src/library.html
@@ -20,7 +20,7 @@
     <include src="./partials/top.html"></include>
     <include src="./partials/footer.html"></include>
     <include src="./partials/modal-team.html"></include>
-    <script type="module" src="./indexLibrary.js"></script>
+    <script type="module" src="./index-library.js"></script>
     
   </body>
 </html>
